server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
nltk==3.8.1
google-generativeai
